% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_slide.R
\name{add_slide}
\alias{add_slide}
\title{Adding a `xaringan` RMarkdown file into the `slides` folder}
\usage{
add_slide(nm = "Untitled.Rmd", template = "default", open = TRUE)
}
\arguments{
\item{nm}{Name of the file}

\item{template}{(default = "default") file path to template file in `whisker` format}

\item{open}{(T/F) should file be opened after creation}
}
\value{
A new `xaringan`-flavored RMarkdown file
}
\description{
Adding a `xaringan` RMarkdown file into the `slides` folder
}
\details{
If a custom template is provided, it should be in the `whisker` format, with
entries taken from the following list:
\itemize{
\item course_shortname (course number)
\item course_name (course name)
\item instructor (instructor)
\item description (description of course)
\item dates (Dates/semester for course)
\item short_repo (short form of github repo, user/repo)
\item course_shortname1 (course number without spaces)
\item github (full github URL)
\item course_url (URL for course homepage)
\item slack (Slack channel for course)
\item twitter (Twitter handle of instructor)
}
}
\examples{
# add_slide('ggplot.Rmd')
}
