% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classDate.R
\name{classDate}
\alias{classDate}
\title{Fill in a sequence of dates for class}
\usage{
classDate(wk, startDate)
}
\arguments{
\item{wk}{The week number (integer >= 1)}
}
\value{
New dates
}
\description{
Fill in a sequence of dates for class
}
