% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_book.R
\name{make_book}
\alias{make_book}
\title{Create the coursenotes using bookdonw}
\usage{
make_book(infile, output_format)
}
\arguments{
\item{infile}{The index file for the course notes}

\item{output_format}{The bookdown output format}
}
\value{
The outcome is a bookdown-based book
}
\description{
Create the coursenotes using bookdonw
}
