% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_hw.R
\name{process_hw}
\alias{process_hw}
\title{Process homework documents to create a mini-website}
\usage{
process_hw(drake_source = here::here("_drake.R"),
  lock_environment = TRUE)
}
\description{
Process homework documents to create a mini-website
}
