% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_slides.R
\name{process_slides}
\alias{process_slides}
\title{Process the slides folder to knit the xaringan slides (using drake) and create the web front-end}
\usage{
process_slides(drake_source = here::here("_drake.R"), pdf = FALSE,
  lock_environment = TRUE)
}
\arguments{
\item{pdf}{Shold PDF versions of the slides be compiled}
}
\description{
Process the slides folder to knit the xaringan slides (using drake) and create the web front-end
}
