% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides2notes.R
\name{slide2note}
\alias{slide2note}
\title{Translate a xaringan slide into a bookdown chapter}
\usage{
slide2note(infile, outdir = "notes", outfile = basename(infile))
}
\arguments{
\item{infile}{Relative path name to the slide RMarkdown fille}

\item{outdir}{Relative path where bookdown chapter will reside}

\item{outfile}{(default = `basename(infile)`) name of output file (without full path)}
}
\value{
A RMarkdown file formatted as a  bookdown chapter
}
\description{
Translate a xaringan slide into a bookdown chapter
}
\examples{
# slide2note('slides/ggplot.Rmd')
}
