% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slide2pdf.R
\name{slide2pdf}
\alias{slide2pdf}
\title{Converting slides to PDF format using chrome_print}
\usage{
slide2pdf(slname)
}
\arguments{
\item{slname}{Either the name of the HTML slides (preferred) or the name of the Rmd file}
}
\description{
Converting slides to PDF format using chrome_print
}
