% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process}
\alias{process}
\title{Build the website, knit the notes, slides and homeworks to create the full website}
\usage{
process(slides = T, notes = T, hw = T, notes_pdf = FALSE)
}
\arguments{
\item{slides}{(T/F) should slides and pages be build}

\item{notes}{(T/F) should course notes be compiled}

\item{hw}{(T/F) should homework be compiled}

\item{notes_pdf}{Should PDF version of the notes be created}
}
\description{
Build the website, knit the notes, slides and homeworks to create the full website
}
