% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slide2pdf.R
\name{slide2pdf}
\alias{slide2pdf}
\title{Converting xaringan slides to PDF}
\usage{
slide2pdf(f)
}
\arguments{
\item{f}{The relative path to the R Markdown file for the xaringan presentation}
}
\value{
The function produces a PDF in the folder where the HTML version of the file resides for web deployment
}
\description{
Converting xaringan slides to PDF
}
\examples{
# slide2pdf('slides/lectures/00-Introduction.Rmd')
}
